requirements.txt
openai==1.35.13
httpx==0.27.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlglot==23.10.0
streamlit==1.38.0